main.py
#!/usr/bin/env python3
"""
Multi-Agent Research System
LLM Engineer Assessment
"""

import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

def agent1_query_processor(query):
    """
    Agent 1: Query Processor
    Decompose user query into 3-5 focused sub-queries
    """
    # TODO: Implement
    pass

def agent2_data_collector(sub_queries):
    """
    Agent 2: Data Collector
    Search and scrape 2-3 web sources per sub-query
    """
    # TODO: Implement
    pass

def agent3_content_analyzer(scraped_data, original_query):
    """
    Agent 3: Content Analyzer
    Synthesize collected data into coherent summary
    """
    # TODO: Implement
    pass

def agent4_report_generator(structured_content, original_query):
    """
    Agent 4: Report Generator
    Create professional PDF report
    """
    # TODO: Implement
    pass

def run_pipeline(query):
    """
    Main pipeline function that orchestrates all agents
    """
    # TODO: Implement pipeline flow
    pass

if __name__ == "__main__":
    # Example usage
    test_query = "Impact of AI in healthcare"
    result = run_pipeline(test_query) 