requirements.txt
groq==0.4.2
requests==2.31.0
beautifulsoup4==4.12.2
reportlab==4.0.4
python-dotenv==1.0.0 